shared types in pong-engine
