+++ packages/pong-engine/src/index.ts
@@ -1,3 +1,4 @@
 export * from './constants';
 export * from './models';
 export * from './game';
+export * from './queue.dto';
